+++ Tests/EventListener/RegisterPeriodicMemoryTimerListenerTest.php
@@ -2,10 +2,12 @@
 
 namespace Gos\Bundle\WebSocketBundle\Tests\EventListener;
 
-use Gos\Bundle\WebSocketBundle\Event\ServerEvent;
+use Gos\Bundle\WebSocketBundle\Event\ServerLaunchedEvent;
 use Gos\Bundle\WebSocketBundle\EventListener\RegisterPeriodicMemoryTimerListener;
 use Gos\Bundle\WebSocketBundle\Server\App\Registry\PeriodicRegistry;
 use PHPUnit\Framework\TestCase;
+use React\EventLoop\LoopInterface;
+use React\Socket\ServerInterface;
 
 class RegisterPeriodicMemoryTimerListenerTest extends TestCase
 {
@@ -30,10 +32,11 @@ class RegisterPeriodicMemoryTimerListenerTest extends TestCase
 
     public function testThePeriodicMemoryTimerIsRegisteredWhenTheServerHasProfilingEnabled(): void
     {
-        $event = $this->createMock(ServerEvent::class);
-        $event->expects($this->once())
-            ->method('isProfiling')
-            ->willReturn(true);
+        $event = new ServerLaunchedEvent(
+            $this->createMock(LoopInterface::class),
+            $this->createMock(ServerInterface::class),
+            true
+        );
 
         $this->listener->registerPeriodicHandler($event);
 
@@ -42,10 +45,11 @@ class RegisterPeriodicMemoryTimerListenerTest extends TestCase
 
     public function testThePeriodicMemoryTimerIsNotRegisteredWhenTheServerHasProfilingDisabled(): void
     {
-        $event = $this->createMock(ServerEvent::class);
-        $event->expects($this->once())
-            ->method('isProfiling')
-            ->willReturn(false);
+        $event = new ServerLaunchedEvent(
+            $this->createMock(LoopInterface::class),
+            $this->createMock(ServerInterface::class),
+            false
+        );
 
         $this->listener->registerPeriodicHandler($event);