+++ Tests/EventListener/RegisterPushHandlersListenerTest.php
@@ -2,7 +2,7 @@
 
 namespace Gos\Bundle\WebSocketBundle\Tests\EventListener;
 
-use Gos\Bundle\WebSocketBundle\Event\ServerEvent;
+use Gos\Bundle\WebSocketBundle\Event\ServerLaunchedEvent;
 use Gos\Bundle\WebSocketBundle\EventListener\RegisterPushHandlersListener;
 use Gos\Bundle\WebSocketBundle\Pusher\ServerPushHandlerInterface;
 use Gos\Bundle\WebSocketBundle\Pusher\ServerPushHandlerRegistry;
@@ -10,6 +10,7 @@ use Gos\Bundle\WebSocketBundle\Server\App\PushableWampServerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use React\EventLoop\LoopInterface;
+use React\Socket\ServerInterface;
 
 class RegisterPushHandlersListenerTest extends TestCase
 {
@@ -53,10 +54,11 @@ class RegisterPushHandlersListenerTest extends TestCase
 
         $this->pushHandlerRegistry->addPushHandler($handler);
 
-        $event = $this->createMock(ServerEvent::class);
-        $event->expects($this->once())
-            ->method('getEventLoop')
-            ->willReturn($loop);
+        $event = new ServerLaunchedEvent(
+            $loop,
+            $this->createMock(ServerInterface::class),
+            false
+        );
 
         $this->listener->registerPushHandlers($event);
     }