+++ Tests/EventListener/StartServerListenerTest.php
@@ -2,7 +2,7 @@
 
 namespace Gos\Bundle\WebSocketBundle\Tests\EventListener;
 
-use Gos\Bundle\WebSocketBundle\Event\ServerEvent;
+use Gos\Bundle\WebSocketBundle\Event\ServerLaunchedEvent;
 use Gos\Bundle\WebSocketBundle\EventListener\StartServerListener;
 use Gos\Bundle\WebSocketBundle\Pusher\ServerPushHandlerRegistry;
 use Gos\Bundle\WebSocketBundle\Server\App\Registry\PeriodicRegistry;
@@ -23,7 +23,7 @@ class StartServerListenerTest extends TestCase
     private $serverPushHandlerRegistry;
 
     /**
-     * @var \Gos\Bundle\WebSocketBundle\EventListener\StartServerListener
+     * @var StartServerListener
      */
     private $listener;
 
@@ -46,14 +46,11 @@ class StartServerListenerTest extends TestCase
         $loop->expects($this->once())
             ->method('addSignal');
 
-        $event = $this->createMock(ServerEvent::class);
-        $event->expects($this->once())
-            ->method('getEventLoop')
-            ->willReturn($loop);
-
-        $event->expects($this->once())
-            ->method('getServer')
-            ->willReturn($this->createMock(ServerInterface::class));
+        $event = new ServerLaunchedEvent(
+            $loop,
+            $this->createMock(ServerInterface::class),
+            false
+        );
 
         $this->listener->bindPnctlEvent($event);
     }